backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
django-filter==23.5
drf-spectacular==0.27.0
boto3==1.34.10
django-storages==1.14.2
django-import-export==3.3.5
pandas==2.1.4
openpyxl==3.1.2
django-modeltranslation==0.18.11
gunicorn==21.2.0
whitenoise==6.6.0

